--- src/shared/ui/dropdown-filter/index.ts
@@ -1,2 +1,0 @@
-export { DropdownFilter } from './DropdownFilter'
-export type { DropdownOption } from './types'
