+++ src/shared/ui/dropdown/index.ts
@@ -1,0 +1,2 @@
+export { Dropdown } from './Dropdown'
+export type { DropdownOption } from './types'
