+++ src/shared/ui/index.ts
@@ -3,4 +3,4 @@ export { RadioButton } from './radio-button'
 export { PercentCircle } from './percent-circle'
 export { Typography } from './typography'
 export { type IconName, Icon } from './icon'
-export { DropdownFilter, type DropdownOption } from './dropdown-filter'
+export { Dropdown, type DropdownOption } from './dropdown'
